requirements.txt
backports.zoneinfo
certifi
charset-normalizer
Django==3.0.10
django-cors-headers==3.7.0
django-environ==0.4.5
djangorestframework==3.11.0
extend-me
httplib2
idna
odoo-rpc-client
pytz
requests
simplejson
six
sqlparse
tzdata
urllib3
